CA-130238: Fixing xsiostat to work with tapdisk3
This commit includes massive changes to xsiostat in order to make it work with
tapdisk3. It also changes hard tabs for 4-space soft tabs, which pretty much
touches all lines in all files.

Currently, tapdisk3 will export all available statistics through /dev/shm.
xsiostat will therefore scan for entries in that directory and report
accordingly. If a VBD is unplugged, xsiostat will detect that and stop
reporting for that particular VBDs. If no VBDs are availalbe for reporting,
it will exit unless the "scan flag" is set.

This version includes a "scan" feature. If set (-s), this will make xsiostat
scan for newly plugged VBDs at every report iteration. It will also prevent
xsiostat from exiting if no VBDs are currently attached or reportable due to
filters.

This version includes a new implementation of "filters". When run, it is
possible to specify a list of DOM IDs and VBD IDs to report on. For example,
consider the following domains and VBDs are plugged:

Dom 1, VBD 51792
Dom 1, VBD 51712
Dom 2, VBD 51792
Dom 2, VBD 51712

Running xsiostat with "-v 51792" will make it report on Dom 1 and 2, but
only on VBDs 51792.
Running xsiostat with "-v 51792 -d 1", will make it report only on Dom 1 for
that specific VBD.
Running xsiostat with "-v 51792 -d 1 -v 51712" will make it report both VBDs
for Dom 1.